requirements.txt
PyYAML==5.1.2
bs4
requests
